ECS - implement GetCourseDetails

+++ src/main/java/com/jcs/education/course/service/repository/CourseRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CourseRepository extends JpaRepository<CourseEntity, Long> {
+public interface CourseRepository extends JpaRepository<CourseEntity, Integer> {
 
     List<CourseEntity> findByCategoryId(Integer categoryId);
